feat: [prd-work-failure-metadata-and-timeout-retry-005] - [Prove timeout handling with mocked and functional coverage]
